Refactor `WikiCorpus` to use distribution logic in `TextCorpus` and move tests for text corpora classes to `test_textcorpus` module.
