+++ gensim/corpora/wikicorpus.py
@@ -20,16 +20,13 @@ module.
 
 import bz2
 import logging
-import multiprocessing
 import re
-import signal
 from xml.etree.cElementTree import \
     iterparse  # LXML isn't faster, so let's go with the built-in solution
 
 from gensim import utils
 # cannot import whole gensim.corpora, because that imports wikicorpus...
-from gensim.corpora.dictionary import Dictionary
-from gensim.corpora.textcorpus import TextCorpus
+from gensim.corpora import textcorpus
 
 logger = logging.getLogger(__name__)
 
@@ -174,7 +171,7 @@ def tokenize(content):
     """
     # TODO maybe ignore tokens with non-latin characters? (no chinese, arabic, russian etc.)
     return [
-        utils.to_unicode(token) for token in utils.tokenize(content, lower=True, errors='ignore')
+        token for token in utils.tokenize(content, lower=True, errors='ignore')
         if 2 <= len(token) <= 15 and not token.startswith('_')
     ]
 
@@ -236,26 +233,7 @@ def extract_pages(f, filter_namespaces=False):
 _extract_pages = extract_pages  # for backward compatibility
 
 
-def process_article(args):
-    """
-    Parse a wikipedia article, returning its content as a list of tokens
-    (utf8-encoded strings).
-    """
-    text, lemmatize, title, pageid = args
-    text = filter_wiki(text)
-    if lemmatize:
-        result = utils.lemmatize(text)
-    else:
-        result = tokenize(text)
-    return result, title, pageid
-
-
-def init_to_ignore_interrupt():
-    """Should only be used when master is prepared to handle termination of child processes."""
-    signal.signal(signal.SIGINT, signal.SIG_IGN)
-
-
-class WikiCorpus(TextCorpus):
+class WikiCorpus(textcorpus.TextCorpus):
     """
     Treat a wikipedia articles dump (\*articles.xml.bz2) as a (read-only) corpus.
 
@@ -267,7 +245,8 @@ class WikiCorpus(TextCorpus):
 
     """
     def __init__(self, fname, processes=None, lemmatize=utils.has_pattern(), dictionary=None,
-                 filter_namespaces=('0',)):
+                 filter_namespaces=('0',), metadata=False, token_filters=None,
+                 character_filters=None):
         """
         Initialize the corpus. Unless a dictionary is provided, this scans the
         corpus once, to determine its vocabulary.
@@ -278,21 +257,38 @@ class WikiCorpus(TextCorpus):
         self.metadata if set to true will ensure that serialize will write out article titles to a pickle file.
 
         """
-        self.fname = fname
         self.filter_namespaces = filter_namespaces
-        self.metadata = False
-        if processes is None:
-            processes = max(1, multiprocessing.cpu_count() - 1)
-        self.processes = processes
-        self.lemmatize = lemmatize
-        if dictionary is None:
-            self.dictionary = Dictionary(self.get_texts())
-        else:
-            self.dictionary = dictionary
+        tokenizer = utils.lemmatize if lemmatize else tokenize
+        if character_filters is None:
+            # no need to lowercase and unicode, because the tokenizer already does that.
+            character_filters = [textcorpus.deaccent, textcorpus.strip_multiple_whitespaces]
+        super(WikiCorpus, self).__init__(fname, dictionary, metadata, character_filters, tokenizer,
+                                         token_filters, processes)
+
+    def getstream(self):
+        """Yield documents from the underlying plain text collection (of one or more files).
+        Each item yielded from this method will be considered a document by subsequent
+        preprocessing methods.
+        """
+        return extract_pages(bz2.BZ2File(self.input), self.filter_namespaces)
+
+    def preprocess_text(self, args):
+        """Parse a wikipedia article, returning its content as a list of tokens
+        (utf8-encoded strings).
+        """
+        title, text, pageid = args
+        text = filter_wiki(text)
+        result = super(WikiCorpus, self).preprocess_text(text)
+        return result, title, pageid
+
+    def should_keep_tokens(self, output):
+        tokens, title, pageid = output
+        return not (
+            len(tokens) < ARTICLE_MIN_WORDS or
+            any(title.startswith(ignore + ':') for ignore in IGNORED_NAMESPACES))
 
     def get_texts(self):
-        """
-        Iterate over the dump, returning text version of each article as a list
+        """Iterate over the dump, returning text version of each article as a list
         of tokens.
 
         Only articles of sufficient length are returned (short articles & redirects
@@ -304,41 +300,9 @@ class WikiCorpus(TextCorpus):
         >>> for vec in wiki_corpus:
         >>>     print(vec)
         """
-        articles, articles_all = 0, 0
-        positions, positions_all = 0, 0
-        texts = \
-            ((text, self.lemmatize, title, pageid)
-             for title, text, pageid
-             in extract_pages(bz2.BZ2File(self.fname), self.filter_namespaces))
-        pool = multiprocessing.Pool(self.processes, init_to_ignore_interrupt)
-
-        try:
-            # process the corpus in smaller chunks of docs, because multiprocessing.Pool
-            # is dumb and would load the entire input into RAM at once...
-            for group in utils.chunkize(texts, chunksize=10 * self.processes, maxsize=1):
-                for tokens, title, pageid in pool.imap(process_article, group):
-                    articles_all += 1
-                    positions_all += len(tokens)
-                    # article redirects and short stubs are pruned here
-                    if len(tokens) < ARTICLE_MIN_WORDS or any(title.startswith(ignore + ':') for ignore in IGNORED_NAMESPACES):
-                        continue
-                    articles += 1
-                    positions += len(tokens)
-                    if self.metadata:
-                        yield (tokens, (pageid, title))
-                    else:
-                        yield tokens
-        except KeyboardInterrupt:
-            logger.warn(
-                "user terminated iteration over Wikipedia corpus after %i documents with %i positions"
-                " (total %i articles, %i positions before pruning articles shorter than %i words)",
-                articles, positions, articles_all, positions_all, ARTICLE_MIN_WORDS)
+        doc_token_stream = self.yield_tokens()
+        if self.metadata:
+            return ((tokens, (pageid, title)) for tokens, title, pageid in doc_token_stream)
         else:
-            logger.info(
-                "finished iterating over Wikipedia corpus of %i documents with %i positions"
-                " (total %i articles, %i positions before pruning articles shorter than %i words)",
-                articles, positions, articles_all, positions_all, ARTICLE_MIN_WORDS)
-            self.length = articles  # cache corpus length
-        finally:
-            pool.terminate()
+            return doc_token_stream
 # endclass WikiCorpus
